COMP: Fix warning
svn-url: http://viewvc.slicer.org/viewvc.cgi/Slicer4?view=revision&revision=17649
git-svn-id: http://svn.slicer.org/Slicer4/trunk@17649 3bd1e089-480b-0410-8dfb-8563597acbee

+++ Libs/MRML/vtkEventBroker.cxx
@@ -548,7 +548,7 @@ vtkObservation *vtkEventBroker::GetNthObservation ( int n )
   ObjectToObservationVectorMap::iterator iter; 
   for(iter=this->SubjectMap.begin(); iter != this->SubjectMap.end(); iter++)  
     {
-    if ( n < count + (int)iter->second.size())
+    if ( static_cast<size_t>(n) < count + iter->second.size())
       {
       return (iter->second[n-count]);
       }
